+++ src/components/Datepicker.vue
@@ -116,6 +116,7 @@ import PickerDay from '~/components/PickerDay.vue'
 import PickerMonth from '~/components/PickerMonth.vue'
 import PickerYear from '~/components/PickerYear.vue'
 import Popup from '~/components/Popup.vue'
+import DisabledDate from '~/utils/DisabledDate'
 
 const validDate = (val) =>
   val === null ||
@@ -283,6 +284,9 @@ export default {
     translation() {
       return this.language
     },
+    disabledDatesClass() {
+      return new DisabledDate(this.utils, this.disabledDates)
+    },
   },
   watch: {
     initialView() {
@@ -292,7 +296,8 @@ export default {
       this.setPageDate()
     },
     value(value) {
-      this.setValue(value)
+      const parsedValue = this.parseValue(value)
+      this.setValue(parsedValue)
     },
   },
   mounted() {
@@ -380,7 +385,14 @@ export default {
      */
     init() {
       if (this.value) {
-        this.setValue(this.value)
+        let parsedValue = this.parseValue(this.value)
+        const isDateDisabled =
+          parsedValue && this.disabledDatesClass.isDateDisabled(parsedValue)
+        if (isDateDisabled) {
+          parsedValue = null
+          this.$emit('input', parsedValue)
+        }
+        this.setValue(parsedValue)
       }
       if (this.isInline) {
         this.setInitialView()
@@ -467,18 +479,25 @@ export default {
      * @param {Date|String|Number|null} date
      */
     setValue(date) {
+      if (!date) {
+        this.setPageDate()
+        this.selectedDate = null
+        return
+      }
+      this.selectedDate = date
+      this.setPageDate(date)
+    },
+    /**
+     * parse a datepicker value from string/number to date
+     * @param {Date|String|Number|null} date
+     */
+    parseValue(date) {
       let dateTemp = date
       if (typeof dateTemp === 'string' || typeof dateTemp === 'number') {
         const parsed = new Date(dateTemp)
         dateTemp = Number.isNaN(parsed.valueOf()) ? null : parsed
       }
-      if (!dateTemp) {
-        this.setPageDate()
-        this.selectedDate = null
-        return
-      }
-      this.selectedDate = dateTemp
-      this.setPageDate(dateTemp)
+      return dateTemp
     },
     /**
      * @param {Object} date